+++ R/plot_trajectory.R
@@ -1,19 +1,24 @@
 #' Create a trajectory plot
 #'
 #' @param data A data frame. Requirements:
-#'   * The structure must be like [market_share].
-#'   * The following columns must have a single value: `sector`, `technology`,
-#'   `region`, `scenario_source`.
-#'   * (Optional) If present, the column `label` is used for data labels.
+#' * The structure must be like [market_share].
+#' * The following columns must have a single value: `sector`, `technology`,
+#' `region`, `scenario_source`.
+#' * (Optional) If present, the column `label` is used for data labels.
 #' @param span_5yr Logical. Use `TRUE` to restrict the time span to 5 years from
 #'   the start year (the default behavior of `qplot_trajectory()`), or use
 #'   `FALSE` to impose no restriction.
 #' @template convert_label
 #' @templateVar fun qplot_trajectory
-#' @templateVar value format_metric
+#' @templateVar value recode_metric_trajectory
 #' @param center_y Logical. Use `TRUE` to center the y-axis around start value
 #'   (the default behavior of `qplot_trajectory()`), or use `FALSE` to not
 #'   center.
+#' @param value_col Character. Name of the column to be used as a value to be
+#'   plotted.
+#' @param perc_y_scale Logical. `FALSE` defaults to using no label conversion.
+#'   Use `TRUE` to convert labels on y-axis to percentage using
+#'   `scales::percent` (the default behavior of `qplot_trajectory()`).
 #'
 #' @seealso [market_share].
 #'
@@ -36,27 +41,38 @@
 #' plot_trajectory(
 #'   data,
 #'   span_5yr = TRUE,
-#'   convert_label = format_metric
+#'   convert_label = recode_metric_trajectory,
+#'   center_y = TRUE,
+#'   value_col = "percentage_of_initial_production_by_scope",
+#'   perc_y_scale = TRUE
 #' )
 plot_trajectory <- function(data,
                             span_5yr = FALSE,
                             convert_label = identity,
-                            center_y = FALSE) {
+                            center_y = FALSE,
+                            value_col = "percentage_of_initial_production_by_scope",
+                            perc_y_scale = FALSE) {
+  lifecycle::deprecate_soft(
+      when = "0.4.0",
+      what = "plot_trajectory(data = 'must be prepped already')",
+      details = api_warning_details("plot_trajectory")
+  )
   env <- list(data = substitute(data))
-  check_plot_trajectory(data, env = env)
+  check_plot_trajectory(data, value_col = value_col, env = env)
 
   data %>%
     prep_trajectory(
       convert_label = convert_label,
       span_5yr = span_5yr,
-      center_y = center_y
+      center_y = center_y,
+      value_col = value_col
     ) %>%
-    plot_trajectory_impl()
+    plot_trajectory_impl(perc_y_scale)
 }
 
-check_plot_trajectory <- function(data, env) {
+check_plot_trajectory <- function(data, value_col, env) {
   stopifnot(is.data.frame(data))
-  crucial <- c(common_crucial_market_share_columns(), "production")
+  crucial <- c(common_crucial_market_share_columns(), value_col)
   hint_if_missing_names(abort_if_missing_names(data, crucial), "market_share")
   abort_if_has_zero_rows(data, env = env)
   enforce_single_value <- c("sector", "technology", "region", "scenario_source")
@@ -71,10 +87,11 @@ check_plot_trajectory <- function(data, env) {
 prep_trajectory <- function(data,
                             convert_label = identity,
                             span_5yr = FALSE,
-                            center_y = FALSE) {
+                            center_y = FALSE,
+                            value_col = "percentage_of_initial_production_by_scope") {
   out <- data %>%
     prep_common() %>%
-    mutate(value = .data$production) %>%
+    mutate(value = !!as.name(value_col)) %>%
     mutate(label = convert_label(.data$label))
 
   if (span_5yr) {
@@ -82,19 +99,6 @@ prep_trajectory <- function(data,
   }
 
   start_year <- min(out$year, na.rm = TRUE)
-  if (!quiet()) {
-    inform(glue(
-      "Normalizing `production` values to {start_year} -- the start year."
-    ))
-  }
-  by <- c("metric", "label")
-  out <- left_join(out, filter(out, .data$year == start_year), by = by) %>%
-    mutate(
-      value = .data$value.x / .data$value.y,
-      year = .data$year.x,
-      technology = .data$technology.x
-    ) %>%
-    rename(sector = .data$sector.x)
 
   cols <- c("year", "metric", "label", "technology", "value", "sector")
   out <- select(out, all_of(cols))
@@ -106,7 +110,9 @@ prep_trajectory <- function(data,
   bind_rows(scenarios, not_scenarios)
 }
 
-plot_trajectory_impl <- function(data) {
+plot_trajectory_impl <- function(data, perc_y_scale = FALSE) {
+  stopifnot(is.logical(perc_y_scale))
+
   p <- ggplot(order_trajectory(data), aes(x = .data$year, y = .data$value))
 
   scenarios <- data %>% filter(is_scenario(metric))
@@ -156,14 +162,21 @@ plot_trajectory_impl <- function(data) {
     xlim = c(min(data$year, na.rm = TRUE), max(data$year, na.rm = TRUE) + 0.7 * year_span)
   )
 
-  p +
+  p <- p +
     coord_cartesian(expand = FALSE, clip = "off") +
     scale_x_continuous(breaks = integer_breaks()) +
     scale_fill_manual(values = scenario_colour(data)$colour) +
     # Calling `scale_fill_manual()` twice is intentional (https://git.io/JnDPc)
     scale_fill_manual(aesthetics = "segment.color", values = line_colours(data)) +
     scale_linetype_manual(values = line_types(data)) +
-    scale_color_manual(values = line_colours(data)) +
+    scale_color_manual(values = line_colours(data))
+
+  if (perc_y_scale) {
+    p <- p +
+      scale_y_continuous(labels = percent)
+  }
+
+  p +
     theme_2dii() +
     theme(axis.line = element_blank(), legend.position = "none") %+replace%
     theme(plot.margin = unit(c(0.5, 4, 0.5, 0.5), "cm"))
@@ -263,9 +276,11 @@ distance_from_start_value <- function(data, value) {
 }
 
 get_area_borders <- function(data, center_y = FALSE) {
-  lower <- 0.9 * min(data$value, na.rm = TRUE)
-  upper <- 1.1 * max(data$value, na.rm = TRUE)
+  lower <- min(data$value, na.rm = TRUE)
+  upper <- max(data$value, na.rm = TRUE)
   span <- upper - lower
+  lower <- lower - 0.1 * span
+  upper <- upper + 0.1 * span
 
   upper_distance <- distance_from_start_value(data, upper) / span
   lower_distance <- distance_from_start_value(data, lower) / span