+++ R/r2dii.plot-package.R
@@ -4,5 +4,9 @@
 # The following block is used by usethis to automatically manage
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start
+#' @importFrom dplyr n
+#' @importFrom lifecycle deprecate_soft
+#' @importFrom scales percent
+#' @importFrom stringr str_count
 ## usethis namespace: end
 NULL