+++ tests/testthat/test-qplot_emission_intensity.R
@@ -1,21 +1,23 @@
 test_that("outputs pretty labels", {
-  data <- filter(sda, sector == "automotive")
+  data <- filter(sda, sector == "cement", region == "global")
   p <- qplot_emission_intensity(data)
 
-  metrics <- unique(p$data$label)
-  pretty <- c("Projected", "Corporate Economy")
+  metrics <- sort(unique(p$data$label))
+  pretty <- sort(
+    c("Projected", "Corporate Economy", "Adjusted Scenario Demo", "Target Demo")
+    )
   expect_equal(metrics, pretty)
 })
 
 test_that("Prints title as expected", {
-  data <- filter(sda, sector == "automotive")
+  data <- filter(sda, sector == "cement", region == "global")
   p <- qplot_emission_intensity(data)
 
   expect_snapshot_output(p$labels$title)
 })
 
 test_that("Prints axis labels as expected", {
-  data <- filter(sda, sector == "automotive")
+  data <- filter(sda, sector == "cement", region == "global")
   p <- qplot_emission_intensity(data)
 
   expect_equal(p$labels$x, "Year")
@@ -26,7 +28,7 @@ test_that("Prints axis labels as expected", {
 })
 
 test_that("Plots a data set with maximum time horizon of 5 years", {
-  data <- filter(sda, sector == "automotive")
+  data <- filter(sda, sector == "cement", region == "global")
   p <- qplot_emission_intensity(data)
   expect_true(diff(year_range(p)) <= 5)
 })