+++ NEWS.md
@@ -1,6 +1,36 @@
+# r2dii.plot 0.3.0
+
+## Breaking changes
+
+* Package is updated to align with the newest version of `r2dii.analysis` 
+(v 0.1.13).
+
+* `qplot_trajectory()` now plots the column 
+'percentage_of_initial_production_by_scope' by default and uses a percentage 
+y-scale.
+
+## New features
+
+* `plot_trajectory()` gains new parameters: `value_col` to indicate which column 
+contains the value to be plotted, and `perc_y_scale` to control whether y-axis 
+should be scaled as percentage.
+
+## Bug fixes and minor enhancements
+
+* `to_title()` keeps the word unchanged if all letters are uppercase, for 
+example: `to_title("SDS") = "SDS"`.
+
+* `qplot_trajectory()` has new labels.
+
+* `qplot_techmix()` has scenario name in label now (#476).
+
+* All plots now start at the start year of 'projected' instead of first common 
+year for all metrics. This allows for bar removal from 
+techmix plots (#390 @Antoine-Lalechere).
+
 # r2dii.plot 0.2.0
 
-New features
+## New features
 
 * `plot_*()` functions gain new parameters to help replicate some of
 `qplot_*()` behaviors.
@@ -23,7 +53,7 @@ New features
     - `scale_fill_r2dii_sector()`
     - `scale_fill_r2dii_tech()`
     
-Enhancements
+## Enhancements
 
 * `plot_trajectory()` now defaults to not center the y-axis around the
 start value.
@@ -35,7 +65,7 @@ visibility (#389 @Antoine-Lalechere).
 
 * The website now links to other r2dii packages (#381 @georgeharris2deg).
 
-Bug fixes
+## Bug fixes
 
 * `plot_trajectory()` now displays only integer years on the x-axis (#403).